requirements.txt
# N8N Workflow Scraper - Production Dependencies
# Version: 1.0 - SCRAPE-001 Setup
# Date: October 9, 2025
# Python: 3.11+

# ============================================================================
# CORE SCRAPING (Essential)
# ============================================================================
playwright==1.40.0              # Modern web scraping with iframe support
beautifulsoup4==4.12.2          # HTML parsing
lxml==4.9.3                     # Fast XML/HTML processing

# ============================================================================
# ASYNC & CONCURRENCY (Essential)
# ============================================================================
aiohttp==3.9.1                  # Async HTTP client
aiofiles==23.2.1                # Async file operations
aiolimiter==1.1.0               # Rate limiting (2 req/sec)
tenacity==8.2.3                 # Retry logic with exponential backoff

# ============================================================================
# DATA PROCESSING (Essential)
# ============================================================================
pandas==2.1.4                   # Dataset operations
numpy==1.26.2                   # Numerical operations
pydantic==2.5.2                 # TypeScript-like validation
pydantic-settings==2.1.0        # Settings management
orjson==3.9.10                  # Fast JSON (3-5x faster than stdlib)

# ============================================================================
# MULTIMODAL PROCESSING (Essential)
# ============================================================================
Pillow==10.1.0                  # Image manipulation
pytesseract==0.3.10             # OCR (text extraction from images)
youtube-transcript-api==0.6.1   # YouTube transcript extraction
yt-dlp==2023.12.30              # Video metadata extraction

# ============================================================================
# DATABASE (Essential - PostgreSQL)
# ============================================================================
sqlalchemy==2.0.23              # ORM for clean database operations
alembic==1.13.1                 # Database migrations
psycopg2-binary==2.9.9          # PostgreSQL adapter for Python

# ============================================================================
# MONITORING & LOGGING (Essential)
# ============================================================================
loguru==0.7.2                   # Beautiful logging
rich==13.7.0                    # Rich terminal output (progress bars)

# ============================================================================
# TESTING (Essential)
# ============================================================================
pytest==7.4.3                   # Test framework
pytest-asyncio==0.21.1          # Async test support
pytest-cov==4.1.0               # Coverage reporting
pytest-mock==3.12.0             # Mocking support
faker==20.1.0                   # Test data generation
psutil==7.1.0                   # Memory usage monitoring for integration tests

# ============================================================================
# EXPORT & SERIALIZATION (Essential)
# ============================================================================
pyarrow==14.0.1                 # Parquet format (fast columnar)
zstandard==0.22.0               # Fast compression

# ============================================================================
# CONFIGURATION (Essential)
# ============================================================================
python-dotenv==1.0.0            # Environment variables
pyyaml==6.0.1                   # YAML config files
click==8.1.7                    # CLI framework

# ============================================================================
# DEVELOPMENT TOOLS (Essential)
# ============================================================================
black==23.12.1                  # Code formatting
ruff==0.1.8                     # Fast linting
mypy==1.7.1                     # Type checking
ipython==8.18.1                 # Interactive shell
pre-commit==3.6.0               # Pre-commit hooks

# ============================================================================
# Total: 33 packages for production-grade scraping system
# Installation time: ~3-5 minutes
# Total size: ~500MB
# ============================================================================
pytest-html==4.1.1   # HTML test reports
